BROKEN: Initial work on implementing blueprints
Not tested. Mostly cleaned up a lot of junk form nipo_api file
